Report-Suitability-Modules/requirements.txt
google-generativeai
google-cloud-documentai
python-dotenv
pydantic>=2
